Struct bip_utracker::request::TrackerRequest

pub struct TrackerRequest<'a> { /* fields omitted */ }

TrackerRequest which encapsulates any request sent to a tracker.

Methods

impl<'a> TrackerRequest<'a>

fn new(
conn_id: u64,
trans_id: u32,
req_type: RequestType<'a>
) -> TrackerRequest<'a>

Create a new TrackerRequest.

fn from_bytes(bytes: &'a [u8]) -> IResult<&'a [u8], TrackerRequest<'a>>

Create a new TrackerRequest from the given bytes.

fn write_bytes<W>(&self, writer: W) -> Result<()> where
W: Write,

Write the TrackerRequest to the given writer.

fn connection_id(&self) -> u64

Connection ID supplied with a request to validate the senders address.

For Connect requests, this will always be equal to 0x41727101980. Therefore, you should not hand out that specific ID to peers that make a connect request.

fn transaction_id(&self) -> u32

Transaction ID supplied with a request to uniquely identify a response.

fn request_type(&self) -> &RequestType

Actual type of request that this TrackerRequest represents.

fn to_owned(&self) -> TrackerRequest<'static>

Create an owned version of the TrackerRequest.