Struct bip_utracker::TrackerClient

pub struct TrackerClient {
    // some fields omitted
}

Tracker client that executes requests asynchronously.

Client will shutdown on drop.

Methods

impl TrackerClient

fn new<H>(bind: SocketAddr, handshaker: H) -> Result<TrackerClient> where H: Handshaker + 'static

Create a new TrackerClient.

fn with_capacity<H>(bind: SocketAddr, handshaker: H, capacity: usize) -> Result<TrackerClient> where H: Handshaker + 'static

Create a new TrackerClient with the given message capacity.

Panics if capacity == usize::max_value().

fn request(&mut self, addr: SocketAddr, request: ClientRequest) -> Option<ClientToken>

Execute an asynchronous request to the given tracker.

If the maximum number of requests are currently in progress, return None.

fn responses(&self) -> ClientResponses

Channel that receives the responses from the trackers queried.