Crate bip_utracker

Library for parsing and writing UDP tracker messages.

Includes a default implementation of a bittorrent UDP tracker client and a customizable trait based implementation of a bittorrent UDP tracker server.

Modules

announce
Messaging primitives for announcing.

contact
Messaging primitives for contact information.

error
Messaging primitives for server errors.

option
Messaging primitives for announce options.

request
Messaging primitives for requests.

response
Messaging primitives for responses.

scrape
Messaging primitives for scraping.

Structs

ClientResponses
Responses received by a specific TrackerClient.

ClientResponsesIter
Iterator over responses received by a specific TrackerClient.

ClientToken
Associates a ClientRequest with a ClientResponse.

TrackerClient
Tracker client that executes requests asynchronously.

TrackerServer
Tracker server that executes responses asynchronously.

Enums

ClientError
Errors occuring as the result of a ClientRequest.

ClientRequest
Request made by the TrackerClient.

ClientResponse
Response received by the TrackerClient.

Traits

Handshaker
Trait for providing a handshaker object with connection information.

ServerHandler
Trait for providing a TrackerServer with methods to service TrackerReqeusts.

Type Definitions

ClientResult
Result type for a ClientRequest.

InfoHash

PeerId

ServerResult
Result type for a ServerHandler.